fixed error in progress reporting formatting

+++ papers/fuzzy-fmt/figs/new-melting.cpp
@@ -352,16 +352,16 @@ data find_energy_new(double temp, double reduced_density, double fv, double gwid
         free_energy += temp*(phi_1 + phi_2 + phi_3)*dV;  //NOTE: temp is actually Boltzman constant times temperature
         //printf("free energy is now... %g\n", free_energy);   //debug
         printf("      finished %.7f%% of the integral\n",
-               100*(i + 1)/(Nl)
-               *(j + 1)/(Nl)
-               *(k + 1)/Nl);
+               100*(i + 1)/double(Nl)
+               *(j + 1)/double(Nl)
+               *(k + 1)/double(Nl));
       }
       printf("   finished %.3f%% of the integral\n",
-             100*(i + 1)/(Nl)
-             *(j + 1)/Nl);
+             100*(i + 1)/double(Nl)
+             *(j + 1)/double(Nl));
     }
     printf("finished %.1f%% of the integral\n",
-           100*(i + 1)/Nl);
+           100*(i + 1)/double(Nl));
   }
 
   printf("free_energy is %g\n", free_energy);
